pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mochi"
version = "0.1.0"
description = "AI-powered mock technical interview coach"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Meadowlark Bradsher"}
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pyttsx3>=2.90",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mochi = "mochi.cli:cli"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
